feat(validations): zod for edit and checkout pet, form default values on edit

+++ src/actions/actions.ts
@@ -3,7 +3,7 @@
 import prisma from '@/lib/db';
 import { PetEssentials } from '@/lib/types';
 import { sleep } from '@/lib/utils';
-import { petFormSchema } from '@/lib/validations';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 import { Pet } from '@prisma/client';
 import { revalidatePath } from 'next/cache';
 
@@ -33,16 +33,24 @@ export async function addPet(petData: unknown) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], petData: PetEssentials) {
+export async function editPet(petId: unknown, petData: unknown) {
   // mock network time
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(petData);
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: petData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -53,14 +61,21 @@ export async function editPet(petId: Pet['id'], petData: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   // mock network time
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: 'Invalid pet ID',
+    };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
